Training/react-langchain/requirements.txt
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
python-dotenv==1.1.1
openai==1.109.1
